document-service/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
ollama